refactor: group agent runtime packages under internal/execution/ (#63)
Implemented via agentic development pipeline.

Moves internal/agent, internal/run, and internal/runstate under a shared
internal/execution/ parent directory to make the agent runtime subsystem
legible in the filesystem. Package declarations are unchanged; only import
paths change. All consumer import paths updated globally (runstate before
run to avoid the substring-prefix collision). Stale boundary-constraint
comments in internal/mcp/client.go and internal/trigger/integration_helpers_test.go
updated. CLAUDE.md Key packages section and settled architectural decisions
updated to reflect the new paths.

Closes #63
