+++ internal/execution/agent/classify_mcp_error_test.go
