+++ internal/execution/agent/metrics_test.go
@@ -7,9 +7,9 @@ import (
 	// Import packages that register their collectors at init time so the
 	// smoke test covers all seven metric families in a single Describe pass.
 	"github.com/prometheus/client_golang/prometheus"
-	_ "github.com/rapp992/gleipnir/internal/agent"
+	_ "github.com/rapp992/gleipnir/internal/execution/agent"
 	"github.com/rapp992/gleipnir/internal/metrics"
-	_ "github.com/rapp992/gleipnir/internal/runstate"
+	_ "github.com/rapp992/gleipnir/internal/execution/runstate"
 	_ "github.com/rapp992/gleipnir/internal/timeout"
 )
