+++ internal/execution/agent/state_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/rapp992/gleipnir/internal/model"
-	"github.com/rapp992/gleipnir/internal/runstate"
+	"github.com/rapp992/gleipnir/internal/execution/runstate"
 	"github.com/rapp992/gleipnir/internal/testutil"
 )
