+++ internal/trigger/cron.go
@@ -16,7 +16,7 @@ import (
 	"github.com/rapp992/gleipnir/internal/db"
 	"github.com/rapp992/gleipnir/internal/model"
 	"github.com/rapp992/gleipnir/internal/policy"
-	"github.com/rapp992/gleipnir/internal/run"
+	"github.com/rapp992/gleipnir/internal/execution/run"
 )
 
 // cronLoopHandle wraps a loop's cancel func so the map stores a pointer we can
