+++ internal/trigger/poll.go
@@ -14,7 +14,7 @@ import (
 	"github.com/rapp992/gleipnir/internal/mcp"
 	"github.com/rapp992/gleipnir/internal/model"
 	"github.com/rapp992/gleipnir/internal/policy"
-	"github.com/rapp992/gleipnir/internal/run"
+	"github.com/rapp992/gleipnir/internal/execution/run"
 )
 
 // toolResolver resolves a single MCP tool by dot-notation name. Used by
